STYLE: Update python scripts to fix RUF046 unnecessary cast to int
https://docs.astral.sh/ruff/rules/unnecessary-cast-to-int/

This commit is done anticipating the update of the ruff-pre-commit hook from v0.9.10 to v0.11.0.

+++ Modules/Loadable/Segmentations/EditorEffects/Python/SegmentEditorEffects/SegmentEditorHollowEffect.py
@@ -99,7 +99,7 @@ class SegmentEditorHollowEffect(AbstractScriptedSegmentEditorEffect):
             selectedSegmentLabelmapSpacing = selectedSegmentLabelmap.GetSpacing()
 
         shellThicknessMM = abs(self.scriptedEffect.doubleParameter("ShellThicknessMm"))
-        shellThicknessPixel = [int(math.floor(shellThicknessMM / selectedSegmentLabelmapSpacing[componentIndex])) for componentIndex in range(3)]
+        shellThicknessPixel = [math.floor(shellThicknessMM / selectedSegmentLabelmapSpacing[componentIndex]) for componentIndex in range(3)]
         return shellThicknessPixel
 
     def updateGUIFromMRML(self):
